Feat: Se implementó el endpoint de Presentación: 1. PresentacionRepository.java 2. PresentacionService.java 3. PresentacionServiceImpl.java 4. PresentacionUtils.java 5. PresentacionController.java 6. Test unitarios para los archivos de los puntos 3, 4 y 5

+++ src/main/java/com/mabc/back_cv/web/services/pantalla/PantallaServiceImpl.java
@@ -74,8 +74,8 @@ public class PantallaServiceImpl implements PantallaService {
 
     @Override
     public void deletePantalla(Long id) {
-        if (id == null) {
-            return;
+        if (id == null || !pantallaRepository.existsById(id)) {
+            throw new IllegalArgumentException("No se encontró una pantalla a eliminar.");
         }
         pantallaRepository.deleteById(id);
     }
